Continued working on spell target system, made firebolt
Widgets for spells to be able to use the spell target system for that specific ability, made the firebolt work if you click the widget and select target location. The visual effect is also replicated. (It doesn't deal any damage though because there is no proper health system yet).  I also added the VFX for the ability and implemented it.