Lowered dragon movement speed and removed collision
Lowered dragon movement speed so you could technically run from it. Removed collision on the dragon for the camera so you can see behind it instead of looking at your own butt when it's chasing you because the collision of the dragon pushes the camera in. That sounds rather funny lol :D